+++ tests/Cooked/BasicUsageSpec.hs
@@ -35,7 +35,7 @@ mintingQuickValue =
   void $
     validateTxSkel $
       txSkelTemplate
-        { txSkelMints = txSkelMintsFromList [(Script.Versioned quickCurrencyPolicy Script.PlutusV3, TxSkelNoRedeemer, "banana", 10)],
+        { txSkelMints = txSkelMintsFromList [(Script.Versioned quickCurrencyPolicy Script.PlutusV3, txSkelEmptyRedeemer, "banana", 10)],
           txSkelOuts = [paysPK alice (quickValue "banana" 10)],
           txSkelSigners = [alice],
           txSkelOpts = def {txOptEnsureMinAda = True}
@@ -57,7 +57,7 @@ consumeAlwaysTrueValidator = do
   void $
     validateTxSkel $
       txSkelTemplate
-        { txSkelIns = Map.fromList [(outref, TxSkelRedeemerForScript ())],
+        { txSkelIns = Map.fromList [(outref, txSkelSomeRedeemer ())],
           txSkelOuts = [paysPK alice (ada 10)],
           txSkelSigners = [alice]
         }