+++ tests/Cooked/ReferenceScriptsSpec.hs
@@ -10,6 +10,7 @@ import Data.Set qualified as Set
 import Ledger.Index qualified as Ledger
 import Optics.Core
 import Plutus.Script.Utils.Ada qualified as Script
+import Plutus.Script.Utils.Scripts qualified as Script
 import Plutus.Script.Utils.Typed qualified as Script
 import Plutus.Script.Utils.V3.Typed.Scripts qualified as Script
 import Plutus.Script.Utils.Value qualified as Script
@@ -99,7 +100,7 @@ checkReferenceScriptOnOref expectedScriptHash refScriptOref = do
   void $
     validateTxSkel
       txSkelTemplate
-        { txSkelIns = Map.singleton oref $ TxSkelRedeemerForScript (),
+        { txSkelIns = Map.singleton oref $ txSkelSomeRedeemer (),
           txSkelInsReference = Set.singleton refScriptOref,
           txSkelSigners = [wallet 1]
         }
@@ -116,8 +117,24 @@ useReferenceScript spendingSubmitter theScript = do
   void $
     validateTxSkel
       txSkelTemplate
-        { txSkelIns = Map.singleton oref $ TxSkelRedeemerForReferenceScript scriptOref (),
+        { txSkelIns = Map.singleton oref $ txSkelSomeRedeemerAndReferenceScript scriptOref (),
           txSkelSigners = [spendingSubmitter]
+        }
+
+referenceMint :: (MonadBlockChain m) => Script.Versioned Script.MintingPolicy -> Script.Versioned Script.MintingPolicy -> Int -> m ()
+referenceMint mp1 mp2 n = do
+  ((!! n) -> mpOutRef) <-
+    validateTxSkel' $
+      txSkelTemplate
+        { txSkelOuts = [paysPK (wallet 1) (ada 2) `withReferenceScript` mp1, paysPK (wallet 1) (ada 10)],
+          txSkelSigners = [wallet 1]
+        }
+  void $
+    validateTxSkel $
+      txSkelTemplate
+        { txSkelMints = txSkelMintsFromList [(mp2, txSkelEmptyRedeemerAndReferenceScript mpOutRef, "banana", 3)],
+          txSkelOuts = [paysPK (wallet 1) (ada 2 <> Script.assetClassValue (Script.AssetClass (Script.scriptCurrencySymbol mp2, "banana")) 3)],
+          txSkelSigners = [wallet 1]
         }
 
 tests :: TestTree
@@ -194,20 +211,20 @@ tests =
                 validateTxSkel'
                   txSkelTemplate
                     { txSkelOuts = [paysScript (alwaysTrueValidator @MockContract) () (ada 42)],
-                      txSkelIns = Map.singleton consumedOref TxSkelNoRedeemer,
+                      txSkelIns = Map.singleton consumedOref txSkelEmptyRedeemer,
                       txSkelSigners = [wallet 1]
                     }
               void $
                 validateTxSkel
                   txSkelTemplate
-                    { txSkelIns = Map.singleton oref (TxSkelRedeemerForReferenceScript consumedOref ()),
+                    { txSkelIns = Map.singleton oref (txSkelSomeRedeemerAndReferenceScript consumedOref ()),
                       txSkelSigners = [wallet 1]
                     },
           testCase "fail from transaction generation for mismatching reference scripts"
             $ testFailsFrom
               def
               ( \case
-                  MCEGenerationError err -> err .==. GenerateTxErrorGeneral "toScriptAndRedeemerData: Wrong reference script hash."
+                  MCEGenerationError err -> err .==. GenerateTxErrorGeneral "toPlutusScriptOrReferenceInput: Wrong reference script hash."
                   _ -> testFailure
               )
               def
@@ -222,10 +239,10 @@ tests =
               void $
                 validateTxSkel
                   txSkelTemplate
-                    { txSkelIns = Map.singleton oref (TxSkelRedeemerForReferenceScript scriptOref ()),
+                    { txSkelIns = Map.singleton oref (txSkelSomeRedeemerAndReferenceScript scriptOref ()),
                       txSkelSigners = [wallet 1]
                     },
-          testCase "phase 1 - fail if using a reference script with 'TxSkelRedeemerForScript'"
+          testCase "phase 1 - fail if using a reference script with 'txSkelSomeRedeemer'"
             $ testFailsFrom
               def
               ( \case
@@ -244,7 +261,7 @@ tests =
               void $
                 validateTxSkel
                   txSkelTemplate
-                    { txSkelIns = Map.singleton oref (TxSkelRedeemerForScript ()),
+                    { txSkelIns = Map.singleton oref (txSkelSomeRedeemer ()),
                       txSkelInsReference = Set.singleton scriptOref,
                       txSkelSigners = [wallet 1]
                     },
@@ -261,5 +278,27 @@ tests =
           testCase "succeed if reference script's requirement is met" $
             testSucceeds def $
               useReferenceScript (wallet 1) (requireSignerValidator $ walletPKHash $ wallet 1)
+        ],
+      testGroup
+        "referencing minting policies"
+        [ testCase "succeed if given a reference minting policy" $
+            testSucceeds def $
+              referenceMint quickCurrencyPolicyV3 quickCurrencyPolicyV3 0,
+          testCase "fail if given the wrong reference minting policy"
+            $ testFails
+              def
+              ( \case
+                  MCEGenerationError (GenerateTxErrorGeneral err) -> err .==. "toPlutusScriptOrReferenceInput: Wrong reference script hash."
+                  _ -> testFailure
+              )
+            $ referenceMint permanentCurrencyPolicyV3 quickCurrencyPolicyV3 0,
+          testCase "fail if referencing the wrong utxo"
+            $ testFails
+              def
+              ( \case
+                  MCEGenerationError (GenerateTxErrorGeneral err) -> err .==. "toPlutusScriptOrReferenceInput: Can't resolve reference script utxo."
+                  _ -> testFailure
+              )
+            $ referenceMint quickCurrencyPolicyV3 quickCurrencyPolicyV3 1
         ]
     ]